+++ packages/filesystem/src/browser/file-dialog/file-dialog-module.ts
@@ -15,9 +15,22 @@
  ********************************************************************************/
 
 import { ContainerModule } from '@theia/core/shared/inversify';
+import { LocationListRenderer, LocationListRendererFactory, LocationListRendererOptions } from '../location';
 import { DefaultFileDialogService, FileDialogService } from './file-dialog-service';
-
+import { FileDialogTreeFiltersRenderer, FileDialogTreeFiltersRendererFactory, FileDialogTreeFiltersRendererOptions } from './file-dialog-tree-filters-renderer';
 export default new ContainerModule(bind => {
     bind(DefaultFileDialogService).toSelf().inSingletonScope();
     bind(FileDialogService).toService(DefaultFileDialogService);
+    bind(LocationListRendererFactory).toFactory(context => (options: LocationListRendererOptions) => {
+        const childContainer = context.container.createChild();
+        childContainer.bind(LocationListRendererOptions).toConstantValue(options);
+        childContainer.bind(LocationListRenderer).toSelf().inSingletonScope();
+        return childContainer.get(LocationListRenderer);
+    });
+    bind(FileDialogTreeFiltersRendererFactory).toFactory(context => (options: FileDialogTreeFiltersRendererOptions) => {
+        const childContainer = context.container.createChild();
+        childContainer.bind(FileDialogTreeFiltersRendererOptions).toConstantValue(options);
+        childContainer.bind(FileDialogTreeFiltersRenderer).toSelf().inSingletonScope();
+        return childContainer.get(FileDialogTreeFiltersRenderer);
+    });
 });